+++ quint/src/runtime/compile.ts
@@ -1,7 +1,7 @@
 /*
  * A compiler to the runtime environment.
  *
- * Igor Konnov, 2022-2023
+ * Igor Konnov, Informal Systems, 2022-2023
  *
  * Copyright (c) Informal Systems 2022-2023. All rights reserved.
  * Licensed under the Apache 2.0.
@@ -9,11 +9,11 @@
  */
 
 import { Either, left, right } from '@sweet-monads/either'
-
 import {
   ErrorMessage, Loc, fromIrErrorMessage, parsePhase1, parsePhase2
 } from '../quintParserFrontend'
 import { Computable, ComputableKind, kindName } from './runtime'
+import { ExecutionListener } from './trace'
 import { QuintModule } from '../quintIr'
 import { CompilerVisitor } from './impl/compilerImpl'
 import { walkDefinition } from '../IRVisitor'
@@ -117,6 +117,7 @@ export function contextNameLookup(
  * @param lookupTable lookup table as produced by the parser
  * @param types type table as produced by the type checker
  * @param mainName the name of the module that may contain state varibles
+ * @param execListener execution listener
  * @param rand the random number generator
  * @returns the compilation context
  */
@@ -126,6 +127,7 @@ export function compile(
   lookupTable: LookupTable,
   types: Map<bigint, TypeScheme>,
   mainName: string,
+  execListener: ExecutionListener,
   rand: () => number): CompilationContext {
   const modulesByName = new Map(modules.map(m => [m.name, m]))
   const lastId = modules.map(m => m.id).sort((a, b) => Number(a - b))[modules.length - 1]
@@ -155,7 +157,7 @@ export function compile(
 
   const main = flattenedModules.find(m => m.name === mainName)
 
-  const visitor = new CompilerVisitor(latestTable, types, rand)
+  const visitor = new CompilerVisitor(latestTable, types, rand, execListener)
   if (main) {
     main.defs.forEach(def => walkDefinition(visitor, def))
   }
@@ -193,6 +195,7 @@ export function compile(
  * @param code text that stores one or several Quint modules,
  *        which should be parseable without any context
  * @param mainName the name of the module that may contain state varibles
+ * @param execListener execution listener
  * @param rand the random number generator
  * @returns the compilation context
  */
@@ -200,6 +203,7 @@ export function compileFromCode(
   idGen: IdGenerator,
   code: string,
   mainName: string,
+  execListener: ExecutionListener,
   rand: () => number): CompilationContext {
   // parse the module text
   return parsePhase1(idGen, code, '<input>')
@@ -216,8 +220,8 @@ export function compileFromCode(
       // collect the errors, but do not stop immediately on error
       const [analysisErrors, analysisResult] = analyzer.getResult()
       const ctx =
-        compile(modules,
-          sourceMap, table, analysisResult.types, mainName, rand)
+        compile(modules, sourceMap, table,
+                analysisResult.types, mainName, execListener, rand)
       const errorLocator = mkErrorMessage(sourceMap)
       return right({
         ...ctx,