+++ quint/src/simulation.ts
@@ -3,24 +3,26 @@
  *
  * Igor Konnov, 2023
  *
- * Copyright (c) Informal Systems 2022. All rights reserved.
+ * Copyright (c) Informal Systems 2023. All rights reserved.
  * Licensed under the Apache 2.0.
  * See License.txt in the project root for license information.
  */
 
-import { Either, left, right } from '@sweet-monads/either'
+import { Either } from '@sweet-monads/either'
+import { Maybe, none } from '@sweet-monads/maybe'
 import { strict as assert } from 'assert'
-import { range } from 'lodash'
-import chalk from 'chalk'
 
 import {
   compileFromCode, contextNameLookup, lastTraceName
 } from './runtime/compile'
 import { ErrorMessage } from './quintParserFrontend'
-import { QuintEx } from './quintIr'
-import { fail, kindName } from './runtime/runtime'
-import { chalkQuintEx } from './repl'
+import { QuintApp, QuintEx } from './quintIr'
+import { Computable, EvalResult } from './runtime/runtime'
+import {
+  ExecutionFrame, newTraceRecorder, noExecutionListener
+} from './runtime/trace'
 import { IdGenerator } from './idGenerator'
+import { verbosity } from './verbosity'
 
 /**
  * Various settings that have to be passed to the simulator to run.
@@ -32,53 +34,36 @@ export interface SimulatorOptions {
   maxSamples: number,
   maxSteps: number,
   rand: () => number,
+  verbosity: number,
 }
+
+export type SimulatorResultStatus =
+  'ok' | 'violation' | 'failure' | 'error'
 
 /**
  * A result returned by the simulator.
  */
 export interface SimulatorResult {
-  status: 'ok' | 'violation',
+  status: SimulatorResultStatus,
   vars: string[],
-  trace: QuintEx,
+  states: QuintEx[],
+  frames: ExecutionFrame[],
+  errors: ErrorMessage[],
+}
+
+function errSimulationResult(status: SimulatorResultStatus,
+                             errors: ErrorMessage[]): SimulatorResult {
+  return {
+    status: 'failure',
+    vars: [],
+    states: [],
+    frames: [],
+    errors: errors,
+  }
 }
 
 /**
- * Print a trace with chalk.
- */
-export function printTrace(out: (line: string) => void, trace: QuintEx) {
-  assert(trace.kind === 'app' && trace.opcode === 'List')
-  const kw = (s: string) => chalk.green(s)
-  const lp = chalk.gray('{')
-  const rp = chalk.gray('}')
-  const eq = chalk.gray('=')
-  const comma = chalk.gray(',')
-  trace.args.forEach((state, index) => {
-    assert(state.kind === 'app'
-           && state.opcode === 'Rec' && state.args.length % 2 === 0)
-
-    out(`${kw('action')} step${index} ${eq} ${kw('all')} ` + lp)
-    range(0, Math.trunc(state.args.length / 2))
-      .forEach(i => {
-        const key = state.args[2 * i]
-        assert(key.kind === 'str')
-        const valueText = chalkQuintEx(state.args[2 * i + 1])
-        out(`  ${key.value}' ${eq} ${valueText}` + comma)
-      })
-    out(rp + '\n')
-  })
-
-  out(`${kw('run')} test ${eq} ` + lp)
-  const testText =
-    range(0, trace.args.length)
-      .map(i => `step${i}`)
-      .reduce((left, name) => `${left}.then(${name})`)
-  out('  ' + testText)
-  out(rp)
-}
-
-/**
- * Run a test suite of a single module.
+ * Execute a run.
  *
  * @param idGen a unique generator of identifiers
  * @param code the source code of the modules
@@ -91,8 +76,7 @@ export function
 compileAndRun(idGen: IdGenerator,
               code: string,
               mainName: string,
-              options: SimulatorOptions):
-                Either<ErrorMessage[], SimulatorResult> {
+              options: SimulatorOptions): SimulatorResult {
   // Once we have 'import from ...' implemented, we should pass
   // a filename instead of the source code (see #8)
 
@@ -116,60 +100,45 @@ module __run__ {
 }
 `
 
-  const ctx = compileFromCode(idGen, wrappedCode, '__run__', options.rand)
+  const recorder = newTraceRecorder(options.verbosity)
+  const ctx = compileFromCode(idGen,
+    wrappedCode, '__run__', recorder, options.rand)
 
   if (ctx.compileErrors.length > 0
       || ctx.syntaxErrors.length > 0
       || ctx.analysisErrors.length > 0) {
-    return left(ctx.syntaxErrors
-                .concat(ctx.analysisErrors)
-                .concat(ctx.compileErrors))
+    const errors =
+      ctx.syntaxErrors
+        .concat(ctx.analysisErrors)
+        .concat(ctx.compileErrors)
+    return errSimulationResult('error', errors)
   } else {
-    // evaluate __runResult
-    const res: Either<ErrorMessage[], boolean> =
+    // evaluate __runResult, which triggers the simulator
+    const res: Either<string, Computable> =
       contextNameLookup(ctx, '__runResult', 'callable')
-        .mapLeft(msg => [{ explanation: msg, locs: [] }] as ErrorMessage[])
-        .map(comp => {
-          const result = comp.eval()
-          if (result.isNone()) {
-            return left(ctx.getRuntimeErrors())
-          }
+    if (res.isLeft()) {
+      const errors = [{ explanation: res.value, locs: [] }] as ErrorMessage[]
+      return errSimulationResult('error', errors)
+    } else {
+      const _ = res.value.eval()
+    }
 
-          const ex = result.unwrap().toQuintEx(idGen)
-          if (ex.kind !== 'bool') {
-            return left([{
-              explanation: 'Expected a Boolean result',
-              locs: [],
-            }])
-          } else {
-            return right(ex.value)
-          }
-        })
-        .join()
+    const frame = recorder.getBestTrace()
+    let status: SimulatorResultStatus = 'failure'
+    if (frame.result.isJust()) {
+      const ex = frame.result.unwrap().toQuintEx(idGen)
+      if (ex.kind === 'bool') {
+        status = ex.value ? 'ok' : 'violation'
+      }
+    }
 
-    return res.map(noViolation => {
-        // evaluate _lastTrace
-        const lastTrace =
-          ctx.values.get(kindName('shadow', lastTraceName)) ?? fail
-        const result = lastTrace.eval()
-        const runtimeErrors = ctx.getRuntimeErrors()
-        if (result.isNone() || runtimeErrors.length > 0) {
-          if (runtimeErrors.length > 0) {
-            return left(runtimeErrors)
-          } else {
-            return left([ {
-              explanation: `${lastTraceName} not found`,
-              locs: [],
-            } ])
-          }
-        } else {
-          return right({
-            status: noViolation ? 'ok' : 'violation',
-            vars: ctx.vars,
-            trace: result.unwrap().toQuintEx(idGen),
-          } as SimulatorResult)
-        }
-      })
-      .join()
+    return {
+      status: status,
+      vars: ctx.vars,
+      states: frame.args.map(e => e.toQuintEx(idGen)),
+      frames: frame.subframes,
+      errors: ctx.getRuntimeErrors(),
+    }
   }
 }
+